\name{fassoc-package}
\alias{fassoc-package}
\alias{fassoc}
\docType{package}
\title{
Variance Component Based Multi-marker Association Test Using Family and Unrelated Data
}
\description{This package contains code for simulation and anlyzing multiple marker family data.}
\details{
\tabular{ll}{
Package: \tab fassoc\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-04-09\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}

\author{
Xuefeng Wang; Nathan Morris 
Maintainer:  Xuefeng Wang <pwxfor@gmail.com>
}

\keyword{ package }