\name{hap}
\alias{hap}
\docType{data}
\title{
Sample Data
}
\description{
This data is sample haplotype data generated from the software cosi.
}
\usage{data(hap)}
\format{
  A data frame with 10000 observations on  502 variables.
}

\source{
Simulated using cosi.
}
\references{
http://www.broad.mit.edu/~sfs/cosi
}
