\name{pos}
\alias{pos}
\docType{data}
\title{
Sample Data
}
\description{
Anotation / postion information for the sample haplotype data.
}
\usage{data(pos)}
\format{
  A data frame with 500 observations on the following 7 variables.
  \describe{
    \item{\code{SNP}}{a numeric vector}
    \item{\code{CHROM}}{a numeric vector}
    \item{\code{CHROM_POS}}{a numeric vector}
    \item{\code{ALLELE1}}{a numeric vector}
    \item{\code{FREQ1}}{a numeric vector}
    \item{\code{ALLELE2}}{a numeric vector}
    \item{\code{FREQ2}}{a numeric vector}
  }
}
\source{
Simulated from the cosi software.
}
\references{
http://www.broad.mit.edu/~sfs/cosi
}
