\name{Tped}
\alias{Tped}
\docType{data}
\title{
Sample Data
}
\description{
Pedigree information
}
\usage{data(Tped)}
\format{
  The format is:
 num [1:847, 1:506] 1 1 1 1 1 2 2 2 3 3 ...
 - attr(*, "dimnames")=List of 2
  ..$ : NULL
  ..$ : chr [1:506] "FID" "IID" "FAT" "MAT" ...
}


\keyword{datasets}
